+++ example/lib/many_own_patterns.dart
@@ -40,7 +40,7 @@ class ManyOwnPatterns extends StatelessWidget {
             fontWeight: FontWeight.w600,
             decoration: TextDecoration.underline,
           ),
-          onTapMatch: (email, index, error) {
+          onTapMatch: (email, match, type, index, error) {
             if (error == null) {
               showSnackBar(context, email ?? "Not found");
             }